Rename ServingConf to ServingConfResource
